CSL-944: extract autosave, manage dependencies.
